Move get_anaconda_version_string to util
Move get_anaconda_version_string from pyanaconda.startup_utils to
pyanaconda.core.util. Otherwise, the Services module will import
a lot of storage objects.

Also simplify the code of this function. It is not necessary to check
the existence of the module pyanaconda.version before its import.
