+++ anaconda.py
@@ -193,7 +193,7 @@ def parse_arguments(argv=None, boot_cmdline=None):
               anaconda options that have been found
     """
     from pyanaconda.argument_parsing import getArgumentParser
-    from pyanaconda.startup_utils import get_anaconda_version_string
+    from pyanaconda.core.util import get_anaconda_version_string
 
     ap = getArgumentParser(get_anaconda_version_string(), boot_cmdline)
 
@@ -316,7 +316,7 @@ if __name__ == "__main__":
             startup_utils.prompt_for_ssh()
             sys.exit(0)
 
-    log.info("%s %s", sys.argv[0], startup_utils.get_anaconda_version_string(build_time_version=True))
+    log.info("%s %s", sys.argv[0], util.get_anaconda_version_string(build_time_version=True))
     if os.path.exists("/tmp/updates"):
         log.info("Using updates in /tmp/updates/ from %s", opts.updateSrc)
